refactor: organize stability tests
